Move the useQueryStringPages hook to the hooks directory
